Move functionality from controller to service
