+++ src/azul/service/source_controller.py
@@ -79,6 +79,3 @@ class SourceController(AppController):
             raise TooManyRequestsError(*e.args)
         else:
             return source_ids
-
-    def _list_public_source_ids(self, catalog: CatalogName) -> set[str]:
-        return self._list_source_ids(catalog, authentication=None)
